--- src/main/java/com/athenhub/stockservice/product/application/dto/RegisterResponse.java
@@ -1,5 +1,0 @@
-package com.athenhub.stockservice.product.application.dto;
-
-import java.util.UUID;
-
-public record RegisterResponse(UUID stockId) {}
