--- src/main/java/com/athenhub/stockservice/product/domain/service/BelongsToValidator.java
@@ -1,12 +1,0 @@
-package com.athenhub.stockservice.product.domain.service;
-
-import com.athenhub.stockservice.product.domain.dto.AccessContext;
-
-public interface BelongsToValidator {
-
-    /**
-     * 사용자가 주어진 컨텍스트의 허브/업체에 소속되어 있는지 확인한다.
-     */
-    boolean belongsTo(AccessContext context);
-
-}
