--- src/main/java/com/athenhub/stockservice/product/domain/service/ProductAccessPermissionChecker.java
@@ -1,9 +1,0 @@
-package com.athenhub.stockservice.product.domain.service;
-
-import com.athenhub.stockservice.product.domain.dto.AccessContext;
-
-import java.util.UUID;
-
-public interface ProductAccessPermissionChecker {
-  boolean canAccess(AccessContext accessContext, UUID productId);
-}
