--- src/main/java/com/athenhub/stockservice/product/infrastructure/MemberRole.java
@@ -1,8 +1,0 @@
-package com.athenhub.stockservice.product.infrastructure;
-
-public enum MemberRole {
-  MASTER_MANAGER,
-  HUB_MANAGER,
-  SHIPPING_AGENT,
-  VENDOR_AGENT
-}
