--- src/main/java/com/athenhub/stockservice/product/infrastructure/MemberStatus.java
@@ -1,8 +1,0 @@
-package com.athenhub.stockservice.product.infrastructure;
-
-public enum MemberStatus {
-  PENDING,
-  REJECTED,
-  ACTIVATED,
-  DEACTIVATED
-}
