--- src/main/java/com/athenhub/stockservice/product/infrastructure/OrganizationType.java
@@ -1,15 +1,0 @@
-package com.athenhub.stockservice.product.infrastructure;
-
-public enum OrganizationType {
-  NONE,
-  HUB,
-  VENDOR;
-
-  public boolean isVendor() {
-    return this == VENDOR;
-  }
-
-  public boolean isHub() {
-    return this == HUB;
-  }
-}
