--- src/main/java/com/athenhub/stockservice/product/infrastructure/client/ProductDetail.java
@@ -1,10 +1,0 @@
-package com.athenhub.stockservice.product.infrastructure.client;
-
-import java.util.UUID;
-
-public record ProductDetail(
-        UUID hubId,
-        UUID vendorId
-) {
-
-}
