+++ src/main/java/com/athenhub/stockservice/stock/application/dto/RegisterResponse.java
@@ -1,0 +1,5 @@
+package com.athenhub.stockservice.stock.application.dto;
+
+import java.util.UUID;
+
+public record RegisterResponse(UUID stockId) {}
