+++ src/main/java/com/athenhub/stockservice/stock/domain/dto/RegisterRequest.java
@@ -1,0 +1,8 @@
+package com.athenhub.stockservice.stock.domain.dto;
+
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+import java.util.UUID;
+
+public record RegisterRequest(
+    @NotNull UUID productId, @NotNull UUID variantId, @Min(1) int quantity) {}
