+++ src/main/java/com/athenhub/stockservice/stock/domain/exception/StockDomainException.java
@@ -1,0 +1,11 @@
+package com.athenhub.stockservice.stock.domain.exception;
+
+import com.athenhub.commoncore.error.AbstractServiceException;
+import com.athenhub.commoncore.error.ErrorCode;
+
+public class StockDomainException extends AbstractServiceException {
+
+  public StockDomainException(ErrorCode errorCode, Object... errorArgs) {
+    super(errorCode, errorArgs);
+  }
+}
