+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/MemberRole.java
@@ -1,0 +1,8 @@
+package com.athenhub.stockservice.stock.infrastructure;
+
+public enum MemberRole {
+  MASTER_MANAGER,
+  HUB_MANAGER,
+  SHIPPING_AGENT,
+  VENDOR_AGENT
+}
