+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/MemberStatus.java
@@ -1,0 +1,8 @@
+package com.athenhub.stockservice.stock.infrastructure;
+
+public enum MemberStatus {
+  PENDING,
+  REJECTED,
+  ACTIVATED,
+  DEACTIVATED
+}
