+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/client/ProductDetail.java
@@ -1,0 +1,10 @@
+package com.athenhub.stockservice.stock.infrastructure.client;
+
+import java.util.UUID;
+
+public record ProductDetail(
+        UUID hubId,
+        UUID vendorId
+) {
+
+}
